+++ examples/factory.rs
@@ -27,7 +27,6 @@ impl FactoryComponent for Counter {
     type Input = CounterMsg;
     type Output = CounterOutput;
     type CommandOutput = ();
-    type Widgets = CounterWidgets;
     type ParentInput = AppMsg;
     type ParentWidget = gtk::Box;
 
@@ -141,7 +140,6 @@ impl SimpleComponent for App {
     type Init = u8;
     type Input = AppMsg;
     type Output = ();
-    type Widgets = AppWidgets;
 
     view! {
         gtk::Window {