+++ examples/log.rs
@@ -16,7 +16,6 @@ impl SimpleComponent for App {
     type Init = u8;
     type Input = Msg;
     type Output = ();
-    type Widgets = AppWidgets;
 
     view! {
         gtk::Window {