+++ relm4-macros/src/lib.rs
@@ -70,7 +70,6 @@ fn gtk_import() -> syn::Path {
 ///     type Init = u8;
 ///     type Input = Msg;
 ///     type Output = ();
-///     type Widgets = AppWidgets;
 ///
 ///     view! {
 ///         gtk::Window {
@@ -143,7 +142,6 @@ fn gtk_import() -> syn::Path {
 /// #     type Init = ();
 /// #     type Input = ();
 /// #     type Output = ();
-/// #     type Widgets = AppWidgets;
 /// #
 /// #     view! {
 /// #         gtk::Window {}
@@ -222,7 +220,6 @@ pub fn component(attributes: TokenStream, input: TokenStream) -> TokenStream {
 ///     type Output = CounterOutput;
 ///     type ParentInput = AppMsg;
 ///     type ParentWidget = gtk::Box;
-///     type Widgets = CounterWidgets;
 ///
 ///     view! {
 ///         root = gtk::Box {