+++ relm4-macros/tests/local_ref_container.rs
@@ -9,7 +9,6 @@ impl SimpleComponent for AppModel {
     type Init = ();
     type Input = ();
     type Output = ();
-    type Widgets = AppWidgets;
 
     view! {
         gtk::Window {