+++ relm4-macros/tests/ui/compile-fail/bad_impl.rs
@@ -8,8 +8,7 @@ impl SimpleComponent for TestComponent {
     type Init = ();
     type Input = ();
     type Output = ();
-    type Widgets = TestWidgets;
-
+    
     view! {
         gtk::Window {}
     }