+++ CHANGES.md
@@ -9,6 +9,7 @@
 + core: Add `gnome_42` and `gnome_43` feature flags
 + core: Implement `RelmContainerExt` for `adw::Squeezer`
 + core: Implement `RelmSetChildExt` for `gtk::WindowHandle`
++ macros: Auto-generate the name of the `Widgets` type if possible
 
 ### Changed